Reduce swap to 8GB
RAM is at least 5x faster than NVMe m2 drive.

That is best case for the NVMe drive, reading and writing a large
block, like a big file.

In small, random access cases, RAM can be thousands of times faster.

Overall, I want to hit OOM errors, then figure out how to batch the
problem so I use less RAM at once, to maintain the fast processing
speed of RAM. I do not want to just throw a terrabyte of swap at the
problem, it is not the fastest solution.