V4: There are no projects
Goals are outcomes, tasks are processes.

Projects are, at most, a label for a set of tasks, but since tasks
already encapsulate the idea of subtasks, projects have no distinct
place. Making projects the interface between goals and tasks is
ambiguous and confusing to me.

Simplifying by dropping projects, and just using tasks.