emacs-packages: fix loading python-mode in org babel blocks
The `mode` option for usepackage defers loading of the package. When
used, the package is not loaded and the config block is not run. The
result is that org-babel doesn't have `python` as a loaded language,
and it cannot run python blocks until `python-mode` is triggered by
opening a python file with a `.py` extention.

I can modify the `mode` option to also open when `.python` files
are opened, which seems to match with the buffer that is opened by
`org-edit-special` bound to `C-c '`. However, that means I must edit a
python block before python blocks can be run after each restart of
emacs.

By removing the `mode` trigger, and just loading python-mode at
startup, I get less surprising behaviour, org-babel blocks containing
python code can be executed immediately.
I also don't need to maintain a separate list of languages I want to
use in org-babel, each language can add itself to org-babel if it is
included in the config.

The downside of removing the `mode` trigger is that the package is
loaded at emacs startup, which makes the emacs startup duration
longer. Given that I tend not to restart emacs often, delaying startup
by a few seconds and avoiding a hang when a package loads for the
first time is a tradeoff I accept.